Render event descriptors in non-TTY mode
This replaces the logs-only view when in non-TTY mode with a view that renders task events in a meaningful way.

Status changes are only printed when they are relevant, and the done state will include a duration.

Logs are printed when added to a task.

Improve calculation of function chain status, avoiding a flip back to pending status when some functions are complete but not others.

Sort by status and then start time for correct ordering.