Add `ocuroot release retry` command
Given a package or release ref, identify failed functions, schedule a new run to retry them and execute it.

Examples of usage added to tests.

Also defines a failed_retried state that indicates a failed function or chain is being retried. This stops the whole release being marked failed.

Fixes #1