Collapse function docs into parent job
This change moves the details of function execution into a single job document, tracking functions as a list.

This is a breaking change for state storage.

Since this was a signficant change, it required a lot of interaction with tests and logs. To help with that, logs are now kept in a logs directory under OCUROOT_HOME. Tests have been updated to kepe hold of these for debugging.

There will need to be some UI fixes in follow on commits before creating a PR.