Show .ocu.star file execution in TUI
Include work items for loading the repo.ocu.star and *.ocu.star files for the repo and release configs respectively.

When loading a release config, show the number of tasks and a count of environments as an aid to debugging (when environments have not been set up, for example).

Fix #66