Add `ocuroot start worker` with dev mode
New command that polls a git repo and runs appropriate push work when new commits come in.

Includes a dev mode that will use the current directory as a git remote for testing and tutorial purposes.

Tested with a script that starts the worker and waits for refs to update.

Created a git package for remote git operations that are used in push and can be used for state in future. Implemented full read and push functionality even though we're only using refs for now.