Add S3 backend
Uses the AWS sdk to access S3 and compatible object stores.

Includes a build tag for integration tests to separate them from the rest of the unit tests. Object store configuration is provided via environment variables (or a .env file at the root of the repo).

Some additional modules have been added to go.mod for something else in progress.