Revert to using git via exec for store
Due to packfile corruption issues with the remote git approach, and less than stellar performance, we're reverting to using the local git client in the filesystem.

There are still some e2e failures, but look to be down to how links are being handled and will be addressed in the next commit.